refactor: lot 5 — réorganiser hooks succession (14 → hooks/) et placement (2 → hooks/)
- Succession : déplacer 14 fichiers use*.ts du root vers features/succession/hooks/
- Placement : déplacer usePlacementSimulatorController.ts et usePlacementExportHandlers.ts vers features/placement/hooks/
- Mettre à jour tous les imports (SuccessionSimulator, index.ts, components/, __tests__/)
- Cohérence avec credit/hooks/, ir/hooks/, per/hooks/
- Vérification : typecheck ✅ + check:arch ✅
