+++ src/features/placement/hooks/usePlacementSimulatorController.ts
